moved base constraints into baseConstraints.py
